The beginnings of an admin interface
I've built out the beginnings of an admin interface which allows the
account with administrative priveleges at this point to add accounts
which belong to various roles. The add account page listens to the
events for the addition of accounts and automatically refreshes to
reflect the new state of the blockchain.