Remove admin UI from nursery page
This functionality is now covered by a dedicated admin section of the
website. A such this commit removes the admin ui from the nuresery page
as outlined in #3. Additionally the page has been updated so that if a
NurseryAdded() event is seen it will update to add the new nursery to
the display.